requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
SQLAlchemy==2.0.35
python-dotenv==1.0.1
httpx==0.27.2
apscheduler==3.10.4
langcodes==3.4.0
jinja2==3.1.4
